14. Zmiana klasy Node na prywatną, przeniesienie klasy Node do klasy IntLinkedList oraz usunięcie pliku Node.java
