+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -1,7 +1,7 @@
 package edu.kis.vh.nursery.list;
 
 public class IntLinkedList {
-    private final int EMPTY = -1;
+    private static final int EMPTY = -1;
     private Node last;
     private int i;
 
@@ -35,6 +35,37 @@ public class IntLinkedList {
         int ret = last.getValue();
         last = last.getPrev();
         return ret;
+    }
+
+    private class Node {
+
+        private final int value;
+        private Node prev, next;
+
+        public Node(int i) {
+            value = i;
+        }
+
+        public int getValue() {
+            return value;
+        }
+
+        public Node getPrev() {
+            return prev;
+        }
+
+        public void setPrev(Node prev) {
+            this.prev = prev;
+        }
+
+        public Node getNext() {
+            return next;
+        }
+
+        public void setNext(Node next) {
+            this.next = next;
+        }
+
     }
 
 }
